API Reference

XLSXasJSON.JSONWorkbook — Type

JSONWorkbook(file::AbstractString; start_line = 1)

start_line of each sheets are considered as JSONPointer for data structure. And each sheets are pared to Array{OrderedDict, 1}

Constructors

JSONWorkbook("Example.xlsx")

Arguments

arguments are applied to all Worksheets within Workbook.

row_oriented : if 'true'(the default) it will look for colum names in '1:1', if false it will look for colum names in 'A:A'

start_line : starting index of position of columnname.

squeeze : squeezes all rows of Worksheet to a singe row.

delim : a String or Regrex that of deliminator for converting single cell to array.

XLSXasJSON.JSONWorksheet — Type

JSONWorksheet

construct 'Array{OrderedDict, 1}' for each row from Worksheet

Constructors

JSONWorksheet("Example.xlsx", "Sheet1")
JSONWorksheet("Example.xlsx", 1)

Arguments

row_oriented : if 'true'(the default) it will look for colum names in '1:1', if false it will look for colum names in 'A:A'

start_line : starting index of position of columnname.

squeeze : squeezes all rows of Worksheet to a singe row.

delim : a String or Regrex that of deliminator for converting single cell to array.

Base.merge — Function

merge(a::JSONWorksheet, b::JSONWorksheet, bykey::AbstractString)

Construct a merged JSONWorksheet from the given JSONWorksheets. If the same Pointer is present in another collection, the value for that key will be the value it has in the last collection listed.